# docs/openapi-auth-spec.yaml
openapi: 3.0.3
info:
  title: Flow Masters API
  description: |
    Enhanced API with dual authentication support for Flow Masters platform.
    
    ## Authentication
    
    This API supports two authentication methods:
    
    1. **Bearer Token (Recommended)**: Use `Authorization: Bearer <token>` header
    2. **Legacy API Key (Deprecated)**: Use `x-api-key: <token>` header
    
    When both headers are present, Bearer token takes precedence.
    
    ## Rate Limiting
    
    - Authenticated requests: 1000/hour
    - Public endpoints: 100/hour per IP
    - Debug endpoints: 50/hour (development only)
    
  version: 2.0.0
  contact:
    name: Flow Masters API Support
    email: api-support@flow-masters.ru
    url: https://flow-masters.ru/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.flow-masters.ru
    description: Production server
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Recommended authentication method using Authorization: Bearer <token>"
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: "Legacy authentication method using x-api-key header (deprecated but supported)"
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error description
      example:
        error: "Invalid API key"
        message: "The provided API key is not valid"
...
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing_auth:
              summary: Missing authentication
              value:
                error: "Missing API key. Use either \"Authorization: Bearer <token>\" or \"x-api-key: <token>\" header"
    ForbiddenError:
      description: Invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid_key:
              summary: Invalid API key
              value:
                error: "Invalid API key"
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/integrations:
    post:
      summary: Execute integration action
      description: Perform CRUD operations on collections via integration API
      tags:
        - Integrations
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
                - collection
              properties:
                action:
                  type: string
                  enum: [find, create, update, delete]
                collection:
                  type: string
                  description: Collection name (e.g., posts, categories)
                data:
                  type: object
                  description: Action-specific data
            examples:
              find_posts:
                summary: Find posts
                value:
                  action: "find"
                  collection: "posts"
                  data:
                    where:
                      status: "published"
                    limit: 5
              create_post:
                summary: Create post
                value:
                  action: "create"
                  collection: "posts"
                  data:
                    title: "New Post"
                    content: "Post content"
                    status: "draft"
      responses:
        '200':
          description: Integration action completed successfully
        '400':
          description: Invalid request data
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...